MVP-3970: Add ConfigAlertModal & refactor alert subscription
